feat: `enrich` Lambda setup
* Resources for the `enrich` Lambda function:
  * S3 bucket to save the enriched results;
  * IAM Role with least-privilege permissions for the Lambda;
  * The Lambda function itself.
* Python code to be run on said Lambda. The code reads the SQS message,
and for every movie it calls the OMDB API to get the movie details.
The new enriched movie JSON is generated with a data warehouse cenario
in mind, making it easier to eventually filter and categorize the
results.
After every movie is enriched, the JSON containing every movie
is sent to the S3 Bucket defined earlier.

Note: The environment variables `OMDB_API_KEY` and
`ARTIFACT_BUCKET_NAME` must be present for the
code to work.